Moved Gauge to ui/components
