+++ public/app/plugins/panel/gauge/GaugePanel.tsx
@@ -1,8 +1,7 @@
 import React, { PureComponent } from 'react';
-import { PanelProps, NullValueMode } from '@grafana/ui';
+import { PanelProps, NullValueMode, Gauge } from '@grafana/ui';
 
-import { getTimeSeriesVMs } from 'app/viz/state/timeSeries';
-import Gauge from 'app/viz/Gauge';
+import { getTimeSeriesVMs } from './timeSeries';
 import { GaugeOptions } from './types';
 
 interface Props extends PanelProps<GaugeOptions> {}
