+++ public/app/plugins/panel/gauge/timeSeries.ts
